feat: drive symbol_at through TS local semantics
- bind TsLocalSemantics during analysis and reuse the stored tables

- resolve symbol_at via hir-js SpanMap + TsLocalSemantics and map top-level locals to module/global symbols

- align decl lowering and hosts with FileKey resolution, fix default import symbol mapping, and refresh tests

+++ typecheck-ts/src/snapshot.rs
@@ -76,5 +76,5 @@ pub struct ProgramSnapshot {
   pub builtin: BuiltinTypes,
   pub next_def: u32,
   pub next_body: u32,
-  pub next_symbol: u32,
+  pub next_symbol: u64,
 }
